Consider cases where real data and/or synthetic data does not exist

+++ src/privateschool/Student.java
@@ -28,7 +28,7 @@ public class Student {
         return dateOfBirth;
     }
 
-    public double getTuitionFees() {
+    public int getTuitionFees() {
         return tuitionFees;
     }
